+++ src/app/prediction.sh
@@ -2,15 +2,15 @@
 
 set -eu
 
-#source "${PWD}/setUpApp.sh"
-. "${PWD}/src/app/setUpApp.sh"
-#source "${DBN_Libraries}/newopts.shlib" "$@"
+# Get absolute path of setUpAPP.sh
+setup=$( cd "$(dirname "$0")" ; pwd )
+. "${setup}/setUpApp.sh"
 . "${DBN_Libraries}/newopts.shlib" "$@"
 
-#this function gets called by opts_ParseArguments when --help is specified
+# This function gets called by opts_ParseArguments when --help is specified
 function usage()
 {
-    #header text
+    # Header text
     echo "
 $log_ToolName: Predict brain age given a brain extracted, MNI registered T1w Image
 
@@ -18,30 +18,32 @@ Usage: $log_ToolName --data=<path to the data folder>
                      --subjects=<path to file with subject IDs>
                      --output=<path to outuput txt file>
                      [--b0=<scanner magnetic field intensity] default=3T
-                     [--model=<path to the .h5 neural network model] default="${DBNDIR}/models/DBN_model.h5"
+                     [--model=<path to the .h5 neural network model>] default="${DBNDIR}/models/DBN_model.h5"
+                     [--linear=<select (non)linear registered image>] default=yes
 
 PARAMETERs are [ ] = optional; < > = user supplied value
 
 Values default to running the example with sample data
 "
-    #automatic argument descriptions
+    # Automatic argument descriptions
     opts_ShowArguments
 }
 
 function main()
 {
-    opts_AddMandatory '--data' 'DATA_FOLDER' 'data folder Path' "a required value; is the path to the study folder holding the preprocessed data" "--dataFolder"
+    opts_AddMandatory '--data' 'DATA_FOLDER' 'data folder path' "a required value; is the path to the study folder holding the preprocessed data" "--dataFolder"
     opts_AddMandatory  '--out' 'OUT_FOLDER' 'path to output directory' "--output" "--outDir" "outputFolder" "outFolder"
     opts_AddOptional  '--subjects' 'SUBJECTS' 'path to file with subject IDs' "an optional value; path to a file with the IDs of the subject to be processed" "default"  "--subject" "--subjectList" "--subjList"
     opts_AddOptional  '--filename' 'OUT_FILE' 'name of the file with brain ages' "default"  "--output" "--outDir" "outputFolder" "outFolder"
     opts_AddOptional  '--b0' 'B0' 'magnetic field intensity' "an optional value; the scanner magnetic field intensity, e.g., 1.5T, 3T, 7T" "3T"
     opts_AddOptional  '--model' 'MODEL' 'path to .h5 model' "an optional value; path to the model.h5 file" "${DBNDIR}/models/DBN_model.h5"
+    opts_AddOptional  '--linear' 'LINEAR' 'select (non)linear registered image' "an optional value; a swtich variable to select either a linear or nonlinear registered image" "yes"
     opts_ParseArguments "$@"
 
-    #display the parsed/default values
+    # Display the parsed/default values
     opts_ShowValues
 
-    #processing code goes here
+    # Processing code goes here
 
     ### Inputs
     #
@@ -79,7 +81,11 @@ function main()
     for subject in $subjList ; do
         echo -e "Subject $subject"
         #subjectImage="${DATA_FOLDER}/${STUDY_NAME}/${PIPELINE}/${subject}/${B0}/MNINonLinear/T1w_brain.nii.gz"
-        subjectImage="${DATA_FOLDER}/${subject}/${B0}/MNINonLinear/T1w_brain.nii.gz"
+        if [ $LINEAR = yes ] ; then
+            subjectImage="${DATA_FOLDER}/${subject}/${B0}/MNILinear/T1w_brain.nii.gz"
+        else
+            subjectImage="${DATA_FOLDER}/${subject}/${B0}/MNINonLinear/T1w_brain.nii.gz"
+        fi
         cp $subjectImage "${SUBJECTS_DIR}/${subject}_${B0}_T1w_brain.nii.gz"
     done
 
@@ -106,5 +112,5 @@ then
 else
     #positional support goes here - just call main with named parameters built from $1, etc
     log_Err_Abort "positional parameter support is not currently implemented"
-    main --data="$1" --out="$2" --subjects="$3" --filename="$4" --b0="$5" --model="$6"
+    main --data="$1" --out="$2" --subjects="$3" --filename="$4" --b0="$5" --model="$6" --linear="$7"
 fi